Refactored shared code to RealmInstance and RealmFileWrapper.
